Move ConsoleSpanExporter
